wip: cross compile cozmo bridge to run on Pi

+++ sim.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
+set +x
+
 SCRIPT_PATH=$(dirname $([ -L $0 ] && echo "$(dirname $0)/$(readlink -n $0)" || echo $0))
 SCRIPT_NAME=`basename ${0}`
 TOPLEVEL=$(cd "$SCRIPT_PATH" && pwd)
 OS_LATEST=$(curl http://ota.pvic.xyz/vic/latestVersion 2> /dev/null)
 NOBUILD=0
 COZMO=0
+PIPACK=0
+TARGETF=".sim"
 
 while [[ $# -gt 0 ]]; do
     case "$1" in
@@ -17,9 +21,15 @@ while [[ $# -gt 0 ]]; do
             COZMO=1
             shift
             ;;
+		-pi|--pi)
+			PIPACK=1
+			COZMO=1
+			shift
+			;;
         -h|--help)
             echo "-nb = don't build victor"
 			echo "-cozmo = cozmo"
+			echo "-pi = build thing for Pi"
             exit 0
             ;;
         *)
@@ -29,14 +39,16 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
+echo "here"
+
 set -e
 
 if [[ $? != "0" || $OS_LATEST == *"404 page"* || $OS_LATEST == "" ]]; then
 	echo "failed to get current WireOS version from server: $OS_LATEST"
 	echo "either curl isn't installed or we can't access the server"
-	if [[ -n $(ls $TOPLEVEL/.sim/dvcbs/*/apq8009-robot-sysfs.img) ]]; then
+	if [[ -n $(ls $TOPLEVEL/${TARGETF}/dvcbs/*/apq8009-robot-sysfs.img) ]]; then
 		# terrible
-		OS_LATEST=$(ls "$TOPLEVEL/.sim/dvcbs" | \grep 3)
+		OS_LATEST=$(ls "$TOPLEVEL/${TARGETF}/dvcbs" | \grep 3)
 		echo "i see $OS_LATEST is already there, using that"
 	else
 		echo "if you're cozmoing, run this script with internet once so required files can be downloaded"
@@ -65,15 +77,31 @@ echo "latest OS: $OS_LATEST"
 
 cd "$TOPLEVEL"
 
-mkdir -p .sim
-cd .sim
+mkdir -p "${TARGETF}"
+cd "${TARGETF}"
 
-if [[ ! -f mprocs ]]; then
-	mkdir -p mprocs
-	wget https://github.com/pvolok/dekit/releases/download/v0.9.6/mprocs-0.9.6-linux-$(uname -m)-musl.tar.gz
-	tar -zxvf mprocs-0.9.6-linux-$(uname -m)-musl.tar.gz
+TARGETARCH="$(uname -m)"
+if [[ $PIPACK == "1" ]]; then
+	TARGETARCH="aarch64"
+fi
+
+if [[ $PIPACK == "1" ]]; then
+	MPROCSDIR="pi"
+	rm -rf "${MPROCSDIR}"
+	mkdir -p "${MPROCSDIR}"
+else
+	MPROCSDIR="."
+fi
+
+if [[ ! -f "${MPROCSDIR}/mprocs" ]]; then
+	cd "${MPROCSDIR}"
+	wget https://github.com/pvolok/dekit/releases/download/v0.9.6/mprocs-0.9.6-linux-${TARGETARCH}-musl.tar.gz
+	tar -zxvf mprocs-0.9.6-linux-${TARGETARCH}-musl.tar.gz
+	rm "mprocs-0.9.6-linux-${TARGETARCH}-musl.tar.gz"
 	chmod +rwx mprocs
 fi
+
+cd "${TOPLEVEL}/${TARGETF}"
 
 if [[ ! -d webots && $COZMO == "0" ]]; then
 	wget https://github.com/cyberbotics/webots/releases/download/R2021b/webots-R2021b-x86-64.tar.bz2
@@ -84,16 +112,23 @@ fi
 
 if [[ $COZMO == "0" ]]; then
 	cd "${TOPLEVEL}/simulator/controllers/vicBodyBridge"
-	WEBOTS_HOME="${TOPLEVEL}/.sim/webots" make
+	WEBOTS_HOME="${TOPLEVEL}/${TARGETF}/webots" make
 	cd "${TOPLEVEL}/simulator/plugins/robot_windows/vicPanel"
-	WEBOTS_HOME="${TOPLEVEL}/.sim/webots" make
+	WEBOTS_HOME="${TOPLEVEL}/${TARGETF}/webots" make
 fi
+
 if [[ $COZMO == "1" ]]; then
 	cd "${TOPLEVEL}/simulator/controllers/vicCozmoBridge"
-	rm -f vicCozmoBridge
-	make
+	make clean
+	if [[ $PIPACK == "1" ]]; then
+		echo "cross-compiling for Pi"
+		CCT="$HOME/.anki/vicos-sdk/dist/5.3.0-r07/prebuilt/bin/arm-oe-linux-gnueabi-clang"
+		CXXT="$HOME/.anki/vicos-sdk/dist/5.3.0-r07/prebuilt/bin/arm-oe-linux-gnueabi-clang++"
+	fi
+	CC="$CCT" CXX="$CXXT" make
 fi
-cd "${TOPLEVEL}/.sim"
+
+cd "${TOPLEVEL}/${TARGETF}"
 
 if [[ ! -d dvcbs ]]; then
 	git clone https://github.com/kercre123/dvcbs -b mountonly --depth=1
@@ -110,7 +145,7 @@ OSPATH="$(pwd)/${OS_LATEST}/edits"
 sudo ./dvcbs.sh -m "$OS_LATEST"
 touch /tmp/isSimMounted
 
-if [[ ! -f "$OSPATH/tmp/isSimMounted" ]]; then
+if [[ ! -f "$OSPATH/tmp/isSimMounted" && $PIPACK == "0" ]]; then
 	sudo rm "$OSPATH/etc/resolv.conf"
 	sudo touch "$OSPATH/etc/resolv.conf"
 	sudo mount --bind /dev "$OSPATH/dev"
@@ -134,6 +169,17 @@ if [[ ! -d /tmp/anki/gateway ]]; then
 	mkdir -p /tmp/anki/gateway
 	openssl req -config tools/sdk/scripts/mac_cert.conf -subj "/C=US/ST=California/L=SF/O=Anki/CN=Vector-Local" -new -x509 -days 1000 -newkey rsa:2048 -nodes -keyout /tmp/anki/gateway/trust.key -out /tmp/anki/gateway/trust.cert
 fi
+
+TARGETPI="${TOPLEVEL}/${TARGETF}/pi"
+if [[ $PIPACK == "1" ]]; then
+	sudo mv "simulator/controllers/vicCozmoBridge/vicCozmoBridge" "${OSPATH}/"
+	sync
+	sudo umount "$OSPATH"
+	cp -r "${TOPLEVEL}/${TARGETF}/dvcbs/${OS_LATEST}" "${TARGETPI}"
+	cp -r simulator/pi "${TARGETPI}/cozmprocs"
+	exit 0
+fi
+
 #cd "${TOPLEVEL}/simulator/mprocs"
 
 MPROCSTOUSE="simulator/mprocs/mprocs.yaml"
@@ -157,7 +203,7 @@ sudo OSPATH="$OSPATH" \
 	WAYLAND_DISPLAY="$WAYLAND_DISPLAY" \
 	XAUTHORITY="${XAUTHORITY:-$HOME/.Xauthority}" \
 	XDG_RUNTIME_DIR="$XDG_RUNTIME_DIR" \
-	"${TOPLEVEL}/.sim/mprocs" -c ${MPROCSTOUSE}
+	"${TOPLEVEL}/${TARGETF}/mprocs" -c ${MPROCSTOUSE}
 #sudo OSPATH=$OSPATH ./simulator/mprocs/run_chroot.sh /anki/bin/vic-engine
 
 sudo umount "$OSPATH/dev"
